util-collections/src/test/java/com/mastfrog/util/collections/ConvertedReadOnlyListTest.java
/*
 * The MIT License
 *
 * Copyright 2020 Mastfrog Technologies.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package com.mastfrog.util.collections;

import java.util.ArrayList;
import java.util.List;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;
import org.junit.Test;

/**
 *
 * @author Tim Boudreau
 */
public class ConvertedReadOnlyListTest {

    @Test
    @SuppressWarnings("element-type-mismatch")
    public void testConversion() {
        List<Integer> origs = new ArrayList<>();
        for (int i = 0; i < 10; i++) {
            origs.add(i);
        }
        List<String> converted = new ConvertedReadOnlyList<>((v) -> {
            return Integer.toString(v.intValue());
        }, origs);
        assertEquals(origs.size(), converted.size());
        for (int i = 0; i < 10; i++) {
            String s = converted.get(i);
            assertEquals(Integer.toString(i), s);
            assertTrue(s + " not claimed present", converted.contains(s));
            assertFalse(converted.contains(i));
        }
    }

}
